+++ Hrms.Core/Policies/DTRPolicies/RegularWorkDayPolicy.cs
@@ -6,7 +6,7 @@ internal class RegularWorkDayPolicy : PayrollPolicyBase<BasicPipelineData, Payro
     public override BasicPipelineData ApplyIfSatisfied(BasicPipelineData line, PayrollContext context)
     {
         var dailyRate = context.Employee.DailyRate;
-        var hourlyRate = dailyRate / context.DailyRecord.ShiftWorkingHour;
+        var hourlyRate = dailyRate / (decimal)context.DailyRecord.ShiftWorkingHour;
         var regularHours = (decimal)context.DailyRecord.RegularNetHours;
         line.Value += hourlyRate * regularHours;
         return line;
